api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
pydantic==2.9.2
